Refactor NameGenerator property initialization
The standard properties for the NameGenerator object had different
default values at different levels in the NameGenerator hierarchy,
starting with the root NameGenerator, which is refined by extension-
specific subtypes. The handling of the default values in this
structure was extremely inconsistent and failing on reload.

1) Deeper levels in the hierarchy had alternate default values for
new instances which are applied on initial load. However, these
defaults different from the XML schema setting.
2) The model browser showed a property as bold if it differed from
the parent value, not if it differed from the default-default or even
from the local-default for the property.
3) A property would serialize if it differed from the default-default.

These differences left a nightmare on load and reload due to the
default-defaults being applied for unsaved properties by the XSD and
custom constructors reinitializing default values that might have been
previously overridden. Nested default settings could both be lost and
reapplied with different combinations. The end result is that reloading
the model would mess with the name generation settings, which would
only be noted the next time an associated target changed and new names
were applied.

The new arrangement matches the displayed settings:
1) All defaults have been removed from the schema file. These are
now applied post-serialization to properties that were not initialized
during deserialization.
2) All defaults have been removed from the DSL model.
3) At the top-level, only non-default values are serialized.
4) At nested levels, only values that differ from the parent are saved.
5) An empty object is saved at a type transition (the first relational
name generator settings, which then has children for column and table
names). This allows us to distinguish between a previously loaded
and a newly created one. The newly created type transition instance is
first assigned its own (possibly alternate) defaults, but the saved
one gets the difference from its context element.

The only difference to the user is that the saved properties will now
reappear. In the XML file you'll see significantly fewer attributes
listed.