Improve ReferenceModeNaming property display
With an object type selected, the reference mode naming extension
properties would show a value of 'Context Default' by default. The
meaning of this could be determined from the dropdown, which would
show the possible values for each of the options and indicate which
of the types (entity name, value type name, reference mode name, or
custom format) were current selected.

For example, if the default was on custom format, do see what this
value was the user had to open the dropdown, see
'Model Default (Custom Format)' on the top row, then look down the
remaining list in the expansion to see Custom Format (<format applied
to selected object type>).

This was change to show the target value with the default, and match
the displayed value to the current value in the dropdown so that there
is no need to open the dropdown to see settings. So, if the context
default is a custom format, we will now show Default (Custom, <format
applied to the selected object type>). There is no longer a need to
open the dropdown or look multiple places in the dropdown to see the
resolved values.

Also, when the ReferenceModeNaming properties are displayed as
extensions with an object type selection there is no problem of
ambiguity if only one ReferenceModeNaming implementation provides
properties, but as soon as there is more than one extension (giving
multiple subtypes of ReferenceModeNaming and DefaultReferenceModeNaming
classes), then these property names clash. The names now have
alternates that are used in there is this occurs. For example, the
ReferencedEntityTypeCustomFormat property in the relational mapping
becomes RelationalReferencedEntityTypeCustomFormat. Obviously, this
is a long property name, so we only return the long form of the name
(starting with Relational) when there is a high likelihood that
another extension will provide the same property.

Note that these are not generally ambiguous when the properties window
is organized by property categories, but there is no way to tell from
a property descriptor whether the properties window is categorized or
simply alphabetized.