Revision caching with git versioning
The toolbox and fact editor colorization settings update when the
package version is updated. Prior to git versioning the fourth field
(Revision) in the assembly version was an increasing number and could
be used on its own to track a version change. With git versioning the
Revision is simply the number of git changes since the last tag, which
reverts to 0 when a new tag is added.

Use both the Build and Revision numbers to get a strictly increasing
comparable number to represent the installed assembly in the user
registry.