Fix name generator init for new refined instance
The NameGenerator.CreateGeneratorsForRefinedInstance function is used
by extensions to create new name generator instances for a newly added
refined instance. This occurs if the extension needs to apply name
generator settings on a per-instance basis in addition to the global
settings. In the case where the name generators have a usage type,
which is a targeted global refinement of the name generation settings,
these refined instances are children of the usage-targeted instances.
However, the properties were being initialized to values matching the
primary generator, not the usage instance (basically, the grandparent
instead of the parent).

Also in this case, an undo of this action was asserting that multiple
properties were being returned to the Uninitialized state. The asserts
are valid in the normal case, but not in the undo case so were updated
accordingly.