Access NORMA schemas for per-user install
The only functional difference between a per-user and full install
of NORMA is that the schema file support is not automatic with the
per user install. This manifests as a lack of IntelliSense and
'Go To Definition' support when NORMA files (.orm files and generated
files) are open in the XML editor. The easiest way to remedy this is
to open the schema catalogs as individual files in the same Visual
Studio session. The XML editor is smart enough to resolve open schemas
and schemas listed in opened catalog files. It is much easier to open
the catalog files than opening dozens of schema files, so we take this
approach.

The catalog files were already modified as part of the initial per-user
support so that they point to the correct file location relative to
the catalog file path, so all we had to do here is open them. The
extension location itself changes with each new version (and versions
can install automatically with per-user installs), so caching the file
location is not an effective approach for the user.

The top-level XML menu that appears when an XML editor is open now has
a 'Load NORMA Schemas' menu item at the bottom. Selecting this item
will load the NORMA catalog and any extension catalogs that are
registered. We also acknowledge the dependency of a number of generated
files on the PLiX schemas and load that catalog as well.

The only caveat here is that the 'Load NORMA Schemas' menu does not
appear unless a .ORM file has been opened during the VS session to force
the NORMA package to load.