Add support for VS2022 installation
Fairly typical upgrade. I had to add a number of new HintPath settings
to find some of the referenced assemblies. Also, the prior build
system had custom code to locate and launch VSCT.exe to compile .vsct
(Visual Studio menu) files. However, this is no longer an option in
64-bit VS2022 build because VSCT.exe is a 32-bit executable. Using the
provided VSCTCompile from the VSSDK extensions instead, but this had
some side effects (the VSSDK extensions always want to build a .vsix
file, which I do not want).

The only other 32/64 issue was an (int)IntPtr cast in a hashcode
routine, which gave an overflow exception with an 8-bit IntPtr. I just
called the existing GetHashCode instead.