Enable external verbalization for SxS VS
Prior to Visual Studio VS2017 NORMA assemblies were easily referenced
by external programs, which could be used to load files, automate
verbalization product and the like. With the move to side-by-side
Visual Studio installation (with a VSIX installer replacing the old
MSI setup) the files are no longer in a fixed location and are much
harder to load externally.

This added a sample project that includes a NORMAResolver.cs that
can be used in projects based on this sample to externally load
NORMA files.

-The VerbalizationManager.LoadFromRegistry entry point was added for
SxS versions.
-The ExtensionModelData.LoadFromRegistry was modified so that passing
the same key twice ignores the second loop. Similar behavior was added
to the VerbalizationManager.LoadFromRegistry.
-A IVerbalizeFilterChildren parameter was added to
VerbalizationManager.Verbalize to control output. Note that this
filter also applies to the passed in root elements.