Role name shapes appear with DisplayAsObjectType
Role name shapes are supposed to be hidden when a FactTypeShape is
marked with DisplayAsObjectType (to display an objectified fact type
as an object type with a mini role box). The role name shapes clear
when this is initially set but were incorrectly reappearing when the
role name was edited.

Also, if DisplayAsObjectType was turned off and the options page
setting for displaying role names was turned off then the shapes would
reappear (DisplayRoleNames.UserDefault was not considered).

With this fix the phantom shapes will disappear on a reload.