Verbalize roles if role children verbalized
Child objects of roles (errors, value and cardinality constraints, and
custom properties) have always verbalized as direct children of the
fact type. This works in some cases, such as errors and constraints,
because the role description is baked into the verbalization. However,
for custom properties (and possibly other extensions) there is no
natural way to distinguish which properties are meant for which roles.

We need to verbalize a header for roles if the role has children, but
not otherwise. This capability was not part of the verbalization
engine, where the IVerbalizeFilterSelf interface was added to block
verbalization early in the process or make verbalization conditional
on the existence of a verbalized child. This is ignored for top-level
verbalization (where Role still implements IRedirectVerbalization to
verbalize as part of the FactType) but a delayed-parent verbalization
with associated indentation is tracked as part of the
BeginVerbalization callback used by every verbalization.

Roles will also be verbalized if there is a role name, which was not
previously shown. The role will be represented by the name of the role
player, but the hyperlink in the verbalization will jump to the role,
not the role player.

Although the roles are listed in the same reading order used for the
fact type verbalization we do not know when the fact type is verbalized
if any particular role will be verbalized. However, for a ring fact
type we need to distinguish the role players by a subscript to avoid
ambiguity. Although the subscripts will be shown in the fact type
verbalization for a derived fact type, to avoid clutter in the basic
representation the choice was made to keep the status quo and not show
them with the fact type in the non-derived ring case. This means the
subscripts are shown only with the verbalized roles and it is left up
to the user to associate these with the role players shown in the fact
type verbalization. The code to switch this and always verbalize with
subscripts is commented out in VerbalizationGenerator.xslt so can be
easily turned on if we change our minds (search for factMockup).