Reverse reading option, display options refactor
The display of two readings (forward and reverse) for binary fact
types is informative but takes a lot of space on the ORM diagram.
This change added a display option to give the user fine-grained
control over when this is shown.

The NORMA options page has several options to control diagram display.
Most of these do not affect the serialized display elements in the
.orm file, but this was not true for role name display and reading
direction indicator. For these settings, loading and saving a file
(with no additional user actions) on a machine with different option
settings would introduce a significant change in the serialized state,
which is a really bad idea for source control. This is especially bad
for role names, which will be remove on a machine with this option off
then restored in default (not very good) positions when a user with
this option enabled loads the file.

In response to this, the following changes where made to these options:
1) A new capacity for setting global display options was added to the
.orm code (singleton DisplayState, with DisplaySetting children
provided by extensions).
2) The 3 display options (role names, reading indicator, and new
reverse reading option) were moved to a new 'Appearance (New File
Defaults)' section. These are applied to the global file options for a
new file but not read otherwise.
3) All of the global properties have analogs on the diagram and fact
type shape, which default to the context object (global for diagram,
diagram for the shape).
4) The diagram now shows a Display Options category with
GlobalDisplayOptions and LocalDisplayOptions properties. These expand
to show the display options for these two categories. Properties with
a ORMDiagramDisplayOption attribute are automatically pulled into
these type descriptors.
5) The global properties on different DisplaySetting instances are
automatically pulled into the diagram properties. These will generally
have the ORMDiagramDisplayOption(true) attribute so they automatically
placed into the global display options. This allows extensions to
easily add other global display option properties.

There were several changes made to accommodate this:
1) The deserialization fixup code did not know that it was loading
from a newly created file. This is now plumbed with the new
IDeserializationFixupListener.LoadingNewFile method, which is
effectively a new constructor parameter without changing existing
code. This allows the new file appearance options to be applied.
2) EditorUtility has a new ConsolidatePropertyDescriptors method for
combining multiple properties into one expandable property.
3) The reading direction indicator is no longer shown on a unary fact
type. These indicate a role order, which is pointless for a unary fact
type.