Fix merge clash on shared base key space
If two link types share a base relationship that does not allow
duplicates then a cross-model drag could end up with a uniqueness
violation in that shared space. The example is dragging a group
between models where an element is included in one model and
excluded in the other.

In this case the existing link (a GroupingElementExclusion) is not
recognized as a match for the imported GroupingElementInclusion, even
though they share a base relationship (GroupingElementRelationship).
The existence of this unmatched instance causes the drag to fail
because the old and new elements share role players.

The fix is two-fold. First, we need to find this element, which is
done by checking for a base role. Second, we need to move any data
from the shared base to the new link instance (of the correct type).
Note that this moving properties and preserving links does not
currently have any NORMA models that need it, but is included for
completeness.