Set the .orm file icon inside Visual Studio
The ShellFileAssociations\.orm entry was pointing to the ormfile file
type, which requires entries in the HKCR registry hive. Side-by-side
installs (VSIX only without MSI) cannot write these registry keys so
were not showing any icon for .orm files in the VS environment.

This affects icons that appear for .orm files inside Visual Studio. It
does not affect the appearance of .orm files in file explorer or
associate this file type with the current NORMA installation.

It would be possible to add a command inside NORMA to let the user
write this registry data, but it would only work until the next
version is installed (possibly automatically), and would require
temporary admin rights to write tot he registry. As is, the user
would need to choose the Visual Studio Version Selector app one time
when opening an .orm file from File Explorer and remember this setting
for the .orm file type.