Custom properties dialog scaling issues
The custom properties editor did not display properly on a scaled
system. Refactoring the code to reduce complexity eliminates the
problem.

This dialog used UserControls for the group and definition editors.
Given that these controls will never be used elsewhere there is very
little reason to do this and the control contents can just as easily
be put into a panel instead of a user control.

The control spacing and container scaled correctly, but the anchoring
for the contained controls pinned them to the original size, not the
larger scaled size, so a large percentage of the area in the dialog
was not used.

This is obviously a WinForms defect. I looked for workarounds and did
not find anything, which doesn't mean none are available. Instead of
tracking a WinForms defect I eliminated the unneeded nested controls.