'Used by derivations' in fact type verbalization
Added the 'Used by derivations' section to fact type Verbalization.
This shows the list of derivations (fact type derivations, subtype
derivations and join paths) that directly reference the selected
fact type. Nested uses (such as those that occur by using a derived
element in a derivation) are not shown. Automatic join paths are
also not included in the displayed list.

The elements are first grouped by fact type/subtype/set constraint/
set comparison constraint role sequence, then sorted alphabetically
and by sequence number as appropriate.

As with the 'Fact Types' list shown with an object type, the display
of this list is limited to the verbalization browser html (it is not
in reports) and can be controlled by the options dialog.

Limiting the scope of this change required a small interface change
to supply the verbalizationTarget (for example, VerbalizationBrowser)
to IVerbalizeCustomChildren.GetCustomChildVerbalizations. This allows
the same filtering to be done before providing a custom child as is
possible with a conditionalMatch in the generated code.

The formatting issues in the Snippets list of the core snippets
documentation file was also fixed. Border radius has been supported
by CSS for a really long time. There is no need for a bad fake.