Enable deontic modality on value constraints
Deontic modality is displayed with both the deontic color and a
leading degree sign before the opening curly bracket. This is the same
pattern used for a deontic cardinality constraint, which is also a
text-only constraint.

Value constraint verbalization has only positive forms, beginning with
'The possible values...'. The deontic form adds the obligation and
removes the possible, so 'It is obligatory that the values...'.

There is currently no validation error to enforce non-decreasing
strength on value restrictions (so, for example, a role value
constraint cannot be more permissive than a restriction on the value
type). If this is added in the future it must consider modality.