Assembly verification in NORMAResolver
The NORMAResolver.cs utility (used to load NORMA assemblies outside
Visual Studio) located the assembly by file name but did not
verify the full .NET name of the assembly before attempting to load
it. On machines with multiple NORMA installations this could pick
the wrong NORMA installation. The NORMA version with assemblies used
to compile the NORMAResolver class will now be chosen more reliably.

The namespace was also adjusted to make using this class more natural.
NORMAResolver is now dropped and is just the class name.