Random overflow failure in ORMCustomTool report
Writing the ORMCustomTool report file was occasionally producing an
arithmetic overflow exception. I finally triggered this in a debugger
under a memory-stressed process and found that a 32-bit uint with the
sign bit set can be converted from an IntPtr but not back to an
IntPtr. So, instead of (IntPtr)((uint)baseIntPtr + offset) we now do
(IntPtr)(int)((uint)baseIntPtr + offset).