Informal descriptions not deleting with no text
Once an InformalDescription was added there was no way to delete it.
The biggest issue with this is the verbalization showed an empty
Informal Description row with nothing in it (except a ".").

There is no attempt to clean these on load or save. Add data, click
off the editor to commit to the model, then delete the description
text and click off again.