Do not serialize empty informal descriptions
Follow up on prior change to remove these when the text becomes empty.

These are an annoyance and not worth the cost of a deserialization
fixup listener to remove, but blocking serialization is cheap.
Existing definitions (only possible before the previous change to
dynamically remove them) will now go away on save, so eventually
they'll naturally clear out of the model.