Add IDomainModelLoading interface
Allow a domain model to implement IDomainModelLoading interface as a
complement to the existing IDomainModelUnloading interface. The intent
of this interface is to allow a domain model to add default elements
from code immediately after the extension is enable. The extension
manager simply adds a prefix for the namespace on the root element,
which tells the extension to load, but there is no mechanism for
adding associated xml before loading.

IDomainModelLoading is called between the xml load completion and the
start of deserialization fixup processing to allow any elements added
here to be processed as if they were originally in the file.

If a domain model implements this interface it will be called when:
1) A new file is created from a template.
2) The domain model does not implement ICustomSerializedDomainModel
so it never has serialized elements.
3) The domain model implements ICustomSerializedDomainModel and the
model was newly added to the file (such as with the extension manager
dialog).

The callback points for IDomainModelUnloading were also centralized
as this callback was missed in some code paths.