Teardown failure rendering global display options
Drawing code and property descriptors can be called during teardown
while the document and Store instance are being destroyed. When this
happens the store's element directory cannot be accessed to retrieve
the global display options. This would be a minor issue if it didn't
also bring down Visual Studio.

Add defensive code to validate the store before retrieving the global
display options.