Deeply mapped object types require concept type
In the oial algorithm an entity type B is deeply mapped into another
entity type A if there is a one-to-one mandatory relationship, with
the B role mandatory. The most common match for this pattern is a
subtype. In determining if a concept type should be created for an
object type any attached subtype pattern (whether identifying or
secondary) forces a concept type to be created on each end. However,
the very similar deeply mapped fact type did not always create an
entity type.

If B in this case is sufficiently sparse (references to it, but no
non-identifying function roles) and A is a subtype then B would not
get a concept type and would effectively disappear from the oial
model. This caused downstream code (like the relational generator)
to crash attempting to resolve B type to something in the oial model.
A minimal model exhibiting the relational crash is:
A subtype of ASup(.id)
B is identified by A (identifying 1-1 mandatory on B)
C(.Id) has a functional reference to B

If this happens it is a good indication that B doesn't add anything
meaningful to the model, but this is for the modeler to decide, not
the absorption algorithm.

This change includes an algorithm update, which will cause the oial
and relational model to regenerate when the file loads. This change
can produce an extra table, but this table can also be marked as
absorbed to return to the previous relational model.