VerbalizationManager.LoadFromRegistry multi dir
LoadFromRegistry fails with multiple verbalization directories.
This is a simple typo (accessing the wrong loop counter).

Also renamed some awkward rule classes in ExternalConstraintLink.
These all deal with the deletion of constraint shapes when they are
not fully connected. Disabling this set of rules enables programmatic
diagram manipulation without the shape deletion. Obviously, with the
rules disabled it is up to the caller to maintain complete shape
connections.