Enable indent-free .orm file serialization
Indented XML in a .orm file is obviously desired if the XML is to
be human-readable or tracked in git. However, if the file is temporary
then there is no reason to write the indentation. For example, if a
file stream is created for code generation then the tab indentation is
just a waste of memory.

Allow external users to specify custom XmlWriterSettings. A default
'new XmlWriterSettings()' will remove indentation. This is used now
for the ORMXmlStream automation object, which provides the serialized
orm model to code generators.