Role not reconnecting with link fact type display
If a link fact type is displayed for an objectification and the fact
type shape sets the DisplayAsObjectType flag to true then the role
player links are removed correctly. However, toggling the property
back on the diagram does not reconnect the original lines.

This occurred because we checked for the presence of a current shape
for the link on the diagram while validating connections for the
fact type shape. (If another shape exists then connector fixup goes
through a different code path in the multi-shape utility.) However,
in this case the other shape connects to a link fact type shape, not
to another shape for this fact type.

Add a caller-controller filter to ORMDiagram.FindShapeForElement to
verify that the found link shape corresponds to the current fact type.