No objectification props with DisplayAsObjectType
Selecting a fact type shape for an objectified fact type pulls three
sets of properties--those shown top level, and those shown for each of
the ObjectifiedFactType and ObjectifyingEntityType expansions. The
primary display in this state is a mashup of properties from various
sources and is distinguished for extension property providers by a
request to provide properties for the obectification instance, not the
fact type or object type.

The property set changes significantly when the DisplayAsObjectType
property is true. In this case we show all of the object type
properties plus a few items from the fact type. Object type extension
properties are bundled with the return set of object type properties.

The problem for extension providers is that we were also requesting
the regular objectification properties in this case. However, these
are very likely to overlap the object type extension properties. Given
that the objectification, fact type and object type are tightly
coupled it is common to provide the same extension properties for
both requests, which means a whole bunch of duplicated properties
when DisplayAsObjectType is true and an extension provider responds
to properties for the 'objectification' object.

Do not request objectification extension properties when
DisplayAsObjectType is true.