Implied objectification allows implied fact type
Although it looks a little odd to attach an implied fact type to an
implicitly objectified fact type it is not blocked in most parts of
the tool. The implied fact types are shown in the model browser and
can be dragged to the design surface, where they give a reasonable
representation of being associated with the implicitly objectified
fact type.

One way to get to this state is to unobjectify a fact type that can
be implicitly objectified. In the core model all this does is change
the Objectification.IsImplied property to true and the shapes then
adjust accordingly. There was explicit code in this transition to
remove the non-proxy links to link fact types, even though these can
be added with no issues in this state. Make the state more consistent
by not removing the role player links in this case.

This is not the best representation, but these shapes are rare and can
only be added explicitly by the user by dragging them out of the model
browser. It doesn't really hurt anything to see them and explicitly
blocking them can negatively impact cases where they are needed
(adding external constraint information to the roles). It is not worth
trying to consistently block this accross the tool, but we do need to
be consistent in letting the be displayed.