Remove 'basedOn' ctor for VerbalizationManager
A constructor to create a clean VerbalizationManager from another
instance is not working as expected. This was added as a non-critical
feature for the "Enable external verbalization for SxS VS" change on
3/3/2022.

The issue is that the IORMToolServices interface (implemented on the
store) already consolidates the default verbalization options provided
by domain model implementations of IVerbalizationOptionProvider and
the CustomOptions collection on the the VerbalizationManager instance
passed to a ModelLoader constructor, which in turn creates the store.

The verbalization engine gets verbalization options from the tool
services, not from the provided verbalization manager, so allowing a
constructor that splits off a clean VerbalizationManager was pointless
because the custom options are already attached to the Store and are
never read from the 'clean' VerbalizationManager instance. Therefore
the 'basedOn' constructor is a bad idea as it does not work as
expected.