FactType child verbalization poorly filtered
Any external verbalization can provide a filter to limit the elements
to verbalize before they are rendered. This filter is passed to the
fact type's IVerbalizeCustomChildren.GetCustomChildVerbalizations
implementation but was not properly applied for some of the child
elements (constraints, instances, anything downstream from role).

Apply the filter to the missed constructs when it is provided.