Model drag/merge failures and xor auto rename
1) During copy/merge operations it is possible for an element to attach
to a group type before the group type attaches to its group. This was
causing a null reference exception.

2) During copy/merge it is possible for the opposite end of a 1-1 to have
an existing 1-1 relationship of the same type with another element. In
this case the newer relationship wins and the old one needs to be
deleted. Note that the DSL-generated code (with the NORMA
modifications) has similar code to check for an existing relationship
on the opposite role player of a 1-1, which protects against the same
duplicate link exception. This should be a rare condition in a
copy/merge and is most likely when there are equivalent elements in
the source model that both match the same identifying merge pattern.

3) Copy/Merge of an exclusive-or constraint renamed it as an exclusion
constraint if the name was automatically generated. Regeneration of
automatic names (ExclusionConstraint1, ExclusiveOrConstraint2, etc.)
is intentional during merge to avoid name duplication. However, for
the exclusion half of an xor constraint, which has a variable default
name based on the associated ExclusiveOrConstraintCoupler link, the
name is being regenerated before the link is merged, so the
ExclusionOrConstraint# pattern is never recreated during merge.

The obsolete generated name also occurred while splitting and
combining an xor constraint, so this was fixed as a general rule to
handle this case as well.