Deontic uniqueness verbalizes without obligation
The 'It is obligatory that...' is missing for the deontic
verbalization of one of the uniqueness patterns (positive two-role
constraint on two binary fact types with forward readings (no front
text) available from the opposite roles). For example, given A r B
and A s C with an external uniqueness on the B and C roles,

For each B and C,
  at most one A r that B and
  s that C.

did not get 'It is obligatory that ...' added for deontic modality.