DisplayState options blocking generator targets
Generator targets are used to support multiple outputs of the same
type during code generation. In this case the names of the files to
output depends on the model, which needs to be loaded to determine
this information. If the .orm file is open in a designer then this
value is retrieved directly through automation. However, if the
generator is run from the solution explorer with 'Run Custom Tool'
and the file is not open then the file load skips non-generative
extensions (all of the diagrams, plus other UI-only elements that are
not serialized).

This is where the display state causes a problem: display states are
materialized in non-generative extensions, but the reference to them
is in the core extension. In retrospect this should have been placed
in a secondary non-generative extension, but there is no way to
transition to that without modifying the file format.

To fix this, create a concrete DisplaySettingPlaceholder element that
will be created if the specified DisplaySetting subtype is not
available. This class will enable the loader, and marking it as
'NonWritten' will stop it from writing back out to the XML.

A second issue here is that this code was being hit running the custom
tool with the designer open, which should not happen. This was caused
by the custom tool setting retrievedGeneratorTargets to false instead
of true, causing the settings to be pulled from both the open designer
and from the non-generative serialized file. Fixing this should save a
few seconds of regeneration time when generator targets are indicated.