Integrating custom UUID reference mode fails
With recently added UUID data type and reference modes (UUID and Uuid)
there is the new possibility of conflicting reference mode names on
load. This code path had never been exercised because the intrinsic
reference modes have not changed since very early in NORMA development.

If there is an existing UUID general reference mode then a new UUID1
value type is create for each matching reference mode pattern.
Furthermore there is no duplicate name error created for these value
types.

The problem is that reference mode validation happens in the
AddIntrinsicElements portion of deserialization, which is very early.
Specifically this is before the object type names are processed to
populated ORMModel.ObjectTypesDictionary, which allows named lookup
of elements. The result was twofold when we created new value types
while processing a reference mode rename:
1) Existing value types were not registered, so lookup failed.
2) New value types were not seen by the loader so were never tracked
by name, and not noticed as a duplicate name.

Moving this loading after name resolution does not work either because
it also missing registration of the new names, so I left the load in
place and used the notifyAdded interface to mark elements as if they
had been loaded during deserialization.

The fix was to add an alternate mechanism to find existing value types.
This lookup is only populated if the file has a reference mode name
conflict.

Two other features:
1) If the reference mode editor is used to change the data type of a
general reference mode and an existing value type of that name is in
use as part of the reference mode pattern, then the data type for the
value type is changed to match. This is associated with the user
gesture, not through a rule.
2) If a custom reference mode and intrinsic reference mode match on
all 4 relevant features (name, kind, format string, data type) then
the custom reference mode is deleted instead of renamed.

So, to 'upgrade' a custom general UUID reference mode.
1) Open the file. You'll have a UUID1 custom reference mode.
2) While the reference mode is associated with one value type (if
you aren't General, switch to it), change the data type to
Numeric: UUID.
3) Change the Kind to Popular.
4) Save the file and open in the XML editor.
5) Replace all occurrences of UUID1 with UUID then save.
6) Close and reopen the file in the ORM designer.
7) Verify that the custom UUID reference mode has been recognized as
redundant and removed.